Move fake vtctldclient so it doesn't create circular import with internal cluster tests
