+++ go/vt/vtadmin/vtctldclient/fakevtctldclient/vtctldclient.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package testutil
+package fakevtctldclient
 
 import (
 	"context"
